Fixed critical errors in fileinterface library for compression handling; added emmax kinship matrix support as first stage of emmax support; partitioned custom program handler methods into individual source files
