+++ Makefile.am
@@ -1,7 +1,7 @@
 SUBDIRS = fileinterface
 AM_CPPFLAGS = -I$(srcdir)/fileinterface -I$(srcdir)/fileinterface/include
 bin_PROGRAMS = statgen-mi.out
-statgen_mi_out_SOURCES = statgen-mi/annotations.h statgen-mi/boost_random.cc statgen-mi/boost_random.h statgen-mi/cargs.cc statgen-mi/cargs.h statgen-mi/config_interpreter.cc statgen-mi/config_interpreter.h statgen-mi/custom_handlers.cc statgen-mi/custom_handlers.h statgen-mi/genotype_draw_methods.h statgen-mi/main.cc statgen-mi/mi_calculator.cc statgen-mi/mi_calculator.h statgen-mi/parameters.cc statgen-mi/parameters.h statgen-mi/prob_vector.cc statgen-mi/prob_vector.h statgen-mi/program_description.h statgen-mi/submission_formatter.cc statgen-mi/submission_formatter.h statgen-mi/transpose_matrix.cc statgen-mi/transpose_matrix.h statgen-mi/utilities.h
+statgen_mi_out_SOURCES = statgen-mi/annotations.h statgen-mi/boost_random.cc statgen-mi/boost_random.h statgen-mi/cargs.cc statgen-mi/cargs.h statgen-mi/config_interpreter.cc statgen-mi/config_interpreter.h statgen-mi/custom_handlers.cc statgen-mi/custom_handlers.h statgen-mi/custom_handlers_emmaxkin.cc statgen-mi/custom_handlers_impute2.cc statgen-mi/custom_handlers_placeholder.cc statgen-mi/custom_handlers_plink.cc statgen-mi/custom_handlers_snptest.cc statgen-mi/custom_handlers_userdefined.cc statgen-mi/genotype_draw_methods.h statgen-mi/main.cc statgen-mi/mi_calculator.cc statgen-mi/mi_calculator.h statgen-mi/parameters.cc statgen-mi/parameters.h statgen-mi/prob_vector.cc statgen-mi/prob_vector.h statgen-mi/program_description.h statgen-mi/submission_formatter.cc statgen-mi/submission_formatter.h statgen-mi/transpose_matrix.cc statgen-mi/transpose_matrix.h statgen-mi/utilities.h
 statgen_mi_out_CXXFLAGS = $(BOOST_CPPFLAGS) -ggdb
 statgen_mi_out_LDADD = fileinterface/lib/libfileinterface.la
 EXTRA_DIST = examples
