+++ examples/mi.program.config
@@ -1,5 +1,3 @@
-#format is: program_handle path_to_program snp_major?
-
 begin snptest
 executable /home/cpalmer/Development/stage/bin/snptest
 snp_major 1
@@ -31,6 +29,23 @@ parameter vif "Variance inflation factor threshold for PLINK analysis"
 parameter out-prefix "Prefix for PLINK per-draw result files"
 parameter out-suffix "Suffix for PLINK per-draw result files"
 end
+
+begin emmaxkin
+executable /home/cpalmer/Development/stage/bin/emmax-kin
+snp_major TRUE
+parameter tfile-prefix "Prefix of PLINK {tped, tfam} per-draw files to be written"
+parameter tfile-suffix "Suffix of PLINK {tped, tfam} per-draw files to be written"
+parameter emmax-pheno-name "Name of phenotype to be analyzed"
+parameter emmax-covar-names "Comma-delimited list of covariates to be analyzed"
+flag ibs "Whether to create IBS kinship matrix"
+flag bn "Whether to create BN kinship matrix"
+end
+
+#begin emmax
+#executable /home/cpalmer/Development/stage/bin/emmax
+#snp_major 1
+#use_existing
+#end
 
 #mach2qtl /path/to/mach2qtl FALSE
