+++ statgen-mi/custom_handlers.cc
@@ -51,411 +51,25 @@ void MI::populate_software_data_handlers(MI::software_input_handler_map &target)
   pm.set_result_parser(plink_process_results);
   pm.set_extra_file_remover(plink_extra_file_remover);
   target["plink"] = pm;
+  ///////////////////EMMAXKIN///////////////////////////
+  pm.clear();
+  pm.set_filename_generator(emmaxkin_filename_generator);
+  pm.set_genotype_handler(emmaxkin_write_tped_line);
+  pm.add_file_handler(emmaxkin_write_tfam_line);
+  pm.add_file_handler(emmaxkin_write_pheno_line);
+  if (!MI::parameters::get_parameter("emmaxkin-emmax-covar-names").empty())
+    pm.add_file_handler(emmaxkin_write_covar_line);
+  pm.set_command_formatter(emmaxkin_format_command);
+  pm.set_result_parser(placeholder_process_results);
+  pm.set_extra_file_remover(placeholder_extra_file_remover);
+  target["emmaxkin"] = pm;
   //////////////////////////////////////////////////////
 }
 
 
 
-
-///////////////////////////////////IMPUTE2////////////////////////////////////////////////////////////////////////
-std::string MI::impute2_filename_generator(unsigned index) {
-  //genotype data first
-  switch (index) {
-  case 0:
-    return MI::parameters::get_parameter("impute2-gen-filename");
-  case 1:
-    return MI::parameters::get_parameter("impute2-sample-filename");
-  default:
-    throw std::domain_error("impute2_filename_generator: invalid index \"" + to_string<unsigned>(index) + "\"");
-  }
-}
-
-void MI::impute2_handle_gen_line(const std::string &line, MI::prob_vector &vec, MI::annotations &annots) {
-  std::string chr = "", rsid = "", pos = "", a1 = "", a2 = "";
-  double p1 = 0.0, p2 = 0.0, p3 = 0.0;
-  std::istringstream strm1(line);
-  if (!(strm1 >> chr >> rsid >> pos >> a1 >> a2))
-    throw std::domain_error("invalid impute2 gen line detected: \"" + line + "\"");
-  while (strm1 >> p1 >> p2 >> p3) {
-    vec.push_back(p1);
-    vec.push_back(p2);
-  }
-  annots.add("chr", chr);
-  annots.add("rsid", rsid);
-  annots.add("pos", pos);
-  annots.add("a1", a1);
-  annots.add("a2", a2);
-}
-
-void MI::impute2_handle_sample_line(const std::string &line, MI::prob_vector &vec, MI::annotations &annot) {
-  //going to have to assume that the sample phenotype/covariate data were preloaded
-  //so just ignore the prob_vector
-  vec.clear();
-  //annot.clear();
-  std::string fid = "", iid = "", value = "", missingness = "";
-  //std::vector<std::string> pheno_names;
-  std::istringstream strm1(line);
-  //if it's a header line
-  if (line.find("ID_1") != std::string::npos) {//first header line
-    if (!(strm1 >> fid >> iid >> missingness))
-      throw std::domain_error("impute2_handle_sample_line: cannot parse first header line \"" + line + "\"");
-    unsigned val_counter = 1;
-    while (strm1 >> value) {
-      //std::cout << "logging pair \"header" << val_counter << "\" \"" << value << "\"" << std::endl;
-      annot.add("header" + to_string<unsigned>(val_counter), value);
-      //pheno_names.push_back(value);
-      ++val_counter;
-    }
-  } else if (line.find("0 0 0") != std::string::npos) {//second header line
-    if (!(strm1 >> fid >> iid >> missingness))
-      throw std::domain_error("impute2_handle_sample_line: cannot parse second header line \"" + line + "\"");
-    unsigned val_counter = 1;
-    while (strm1 >> value) {
-      //std::cout << "logging type \"" << annot.get("header" + to_string<unsigned>(val_counter),0) << "_type\" \"" << value << "\"" << std::endl;
-      annot.add(annot.get("header" + to_string<unsigned>(val_counter), 0) + "_type", value);
-      ++val_counter;
-    }
-  } else {//data line
-    if (!(strm1 >> fid >> iid >> missingness))
-      throw std::domain_error("impute2_handle_sample_line: cannot parse data line \"" + line + "\"");
-    annot.add("fid", fid);
-    annot.add("iid", iid);
-    annot.add("missingness", missingness);
-    unsigned val_counter = 1;
-    while (strm1 >> value) {
-      //std::cout << "logging phenotype data under value \"" + annot.get("header" + to_string<unsigned>(val_counter), 0) << "\"" << std::endl;
-      annot.add(annot.get("header" + to_string<unsigned>(val_counter), 0), value);
-      ++val_counter;
-    }
-  }
-}
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-
-///////////////////////////////SNPTEST////////////////////////////////////////////////////////////////////////////
-
-std::string MI::snptest_filename_generator(unsigned index, unsigned draw) {
-  //genotype data first
-  switch (index) {
-  case 0:
-    return MI::parameters::get_parameter("snptest-gen-prefix") + to_string<unsigned>(draw)
-      + MI::parameters::get_parameter("snptest-gen-suffix");
-  case 1:
-    return MI::parameters::get_parameter("snptest-sample-prefix") + to_string<unsigned>(draw)
-      + MI::parameters::get_parameter("snptest-sample-suffix");
-  case 2:
-    return MI::parameters::get_parameter("snptest-output-prefix") + to_string<unsigned>(draw)
-      + MI::parameters::get_parameter("snptest-output-suffix");
-  default:
-    throw std::domain_error("snptest_filename_generator: invalid index \"" + to_string<unsigned>(index) + "\"");
-  }
-}
- 
-std::string MI::snptest_write_gen_line(const MI::prob_vector &vec, const MI::annotations &annot, unsigned index) {
-  //need chr, rsid, pos, a1, a2 from annot
-  std::ostringstream o;
-  o << annot.get("chr", index) << ' ' << annot.get("rsid", index) << ' '
-    << annot.get("pos", index) << ' ' << annot.get("a1", index) << ' '
-    << annot.get("a2", index);
-  double p1 = 0.0, p2 = 0.0;//, p3 = 0.0;
-  unsigned res = 0;
-  for (unsigned i = 0; i < vec.size(); i += 2) {
-    p1 = vec.at(i);
-    p2 = vec.at(i+1);
-    /*p3 = 1.0 - p1 - p2;
-    if (p3 > 1.0) p3 = 1.0;
-    if (p3 < 0.0) p3 = 0.0;*/
-    res = standard_genotype_draw(p1, p2);
-    if (res == 1) {
-      o << " 1 0 0";
-    } else if (res == 2) {
-      o << " 0 1 0";
-    } else {
-      o << " 0 0 1";
-    }
-  }
-  return o.str();
-}
-
-std::string MI::snptest_write_sample_line(const MI::annotations &annot) {
-  std::string requested_covariates = MI::parameters::get_parameter("snptest-covar");
-  while (requested_covariates.find(",") != std::string::npos)
-    requested_covariates[requested_covariates.find(",")] = ' ';
-  std::string requested_phenotype = MI::parameters::get_parameter("snptest-pheno");
-  std::ostringstream o;
-  o << "ID_1 ID_2 missing " << requested_phenotype << ' ' << requested_covariates << '\n';
-  o << "0 0 0";// P";
-  std::vector<std::string> values;
-  //sigh. have to determine the types
-  bool trait_is_01 = true, trait_is_12 = true, trait_is_integral = true;
-  if (annot.get(requested_phenotype + "_type", 0).empty() || MI::parameters::get_flag("snptest-autocorrect-traits")) {
-    values = annot.get(requested_phenotype);
-    if (values.empty()) throw std::domain_error("snptest: phenotype \"" + MI::parameters::get_parameter("snptest-pheno") + "\" not found");
-
-    snptest_test_trait_type(values, trait_is_01, trait_is_12, trait_is_integral);
-    if (trait_is_01 || trait_is_12) {
-      o << " B";
-    } else if (trait_is_12) {
-      o << " P";
-    }
-  } else {
-    o << ' ' << annot.get(requested_phenotype + "_type", 0);
-  }
-
-  bool covar_is_01 = true, covar_is_12 = true, covar_is_integral = true;
-  std::istringstream strm1(requested_covariates);
-  std::string current_covariate = "";
-  std::vector<std::string> all_covariates;
-  while (strm1 >> current_covariate) {
-    all_covariates.push_back(current_covariate);
-    if (annot.get(current_covariate + "_type", 0).empty() || MI::parameters::get_flag("snptest-autocorrect-traits")) {
-      values = annot.get(current_covariate);
-      if (values.empty()) throw std::domain_error("snptest: covariate \"" + current_covariate + "\" not found");
-      snptest_test_trait_type(values, covar_is_01, covar_is_12, covar_is_integral);
-      if (covar_is_integral) {
-	o << " D";
-      } else {
-	o << " C";
-      }
-    } else {
-      o << ' ' << annot.get(current_covariate + "_type", 0);
-    }
-  }
-  o << '\n';
-  //now actually write the values
-  for (unsigned i = 0; i < annot.get_count_at("fid"); ++i) {
-    o << annot.get("fid", i) << ' ' << annot.get("iid", i) << " 0 ";
-    if (trait_is_12) {
-      o << (!annot.get(requested_phenotype, i).compare("2") ? "1" : (!annot.get(requested_phenotype, i).compare("1") ? "0" : annot.get(requested_phenotype, i)));
-    } else {
-      o << annot.get(requested_phenotype, i);
-    }
-    for (std::vector<std::string>::const_iterator iter = all_covariates.begin(); iter != all_covariates.end(); ++iter) {
-      o << ' ' << annot.get(*iter, i);
-    }
-    if (i != annot.get_count_at("fid")-1) o << '\n';
-  }
-  return o.str();
-}
-
-void MI::snptest_test_trait_type(std::vector<std::string> &values, bool &trait_is_01, bool &trait_is_12, bool &trait_is_integral) {
-  trait_is_01 = trait_is_12 = trait_is_integral = true;
-  for (std::vector<std::string>::iterator iter = values.begin(); iter != values.end(); ++iter) {
-    if (iter->compare("-9") || cicompare(*iter, "na")) {
-      *iter = "NA";
-      continue;
-    }
-    if (iter->compare("0") && iter->compare("1")) {
-      trait_is_01 = false;
-    }
-    if (iter->compare("1") && iter->compare("2")) {
-      trait_is_12 = false;
-    }
-    if (iter->find(".") != std::string::npos &&
-	iter->find("-") != std::string::npos) {
-      trait_is_integral = false;
-    }
-  }
-}
-
-std::string MI::snptest_format_command(std::string(*filename_generator)(unsigned, unsigned), unsigned draw) {
-  std::string command = "";
-  command += " -data " + filename_generator(0, draw) + " " + filename_generator(1, draw)
-    + " -frequentist " + MI::parameters::get_parameter("snptest-frequentist") + " -pheno "
-    + MI::parameters::get_parameter("snptest-pheno");
-  if (!MI::parameters::get_parameter("snptest-covar").empty())
-    command += " -cov_all";
-  command += " -method " + MI::parameters::get_parameter("snptest-method") + " -o "
-    + filename_generator(2, draw);
-  return command;
-}
-
-bool MI::snptest_process_results(const std::string &result_line,
-				 std::string &rsid,
-				 double &beta,
-				 double &stderr,
-				 std::string &effect_allele) {
-  if (result_line.find("#") != std::string::npos ||
-      result_line.find_first_not_of(" \t") == std::string::npos ||
-      result_line.find("alternate_ids") != std::string::npos) return false;
-  std::istringstream strm1(result_line);
-  std::string catcher = "";
-  //std::cout << "provided line \"" << result_line << "\"" << std::endl;
-  if (!(strm1 >> catcher >> rsid))
-    throw std::domain_error("snptest: insufficient entries in results line \"" + result_line + "\"");
-  for (unsigned i = 0; i < 4; ++i) {
-    if (!(strm1 >> effect_allele))
-      throw std::domain_error("snptest: insufficient entries in results line \"" + result_line + "\"");
-  }
-  for (unsigned i = 0; i < 37; ++i) strm1 >> catcher;
-  if (!(strm1 >> beta >> stderr)) {
-    //throw std::domain_error("snptest: insufficient entries in results line
-    beta = stderr = 1.0 / 0.0;
-  }
-  //std::cout << "got values " << rsid << ' ' << effect_allele << ' ' << beta << ' ' << stderr << std::endl;
-  return true;
-}
-
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-/////////////////////////////////////////////////SNPTEST//////////////////////////////////////////////////////////
-std::string MI::plink_filename_generator(unsigned index, unsigned draw) {
-  switch (index) {
-  case 0:
-    return MI::parameters::get_parameter("plink-bfile-prefix") + to_string<unsigned>(draw)
-      + MI::parameters::get_parameter("plink-bfile-suffix") + ".bed";
-  case 1:
-    return MI::parameters::get_parameter("plink-bfile-prefix") + to_string<unsigned>(draw)
-      + MI::parameters::get_parameter("plink-bfile-suffix") + ".bim";
-  case 2:
-    return MI::parameters::get_parameter("plink-bfile-prefix") + to_string<unsigned>(draw)
-      + MI::parameters::get_parameter("plink-bfile-suffix") + ".fam";
-  case 3:
-    return MI::parameters::get_parameter("plink-out-prefix") + to_string<unsigned>(draw)
-      + MI::parameters::get_parameter("plink-out-suffix") + ".pheno";
-  case 4:
-    return MI::parameters::get_parameter("plink-out-prefix") + to_string<unsigned>(draw)
-      + MI::parameters::get_parameter("plink-out-suffix")
-      + (MI::parameters::get_flag("plink-linear") ? ".assoc.linear" : ".assoc.logistic");
-  default:
-    throw std::domain_error("plink-filename-generator: invalid index \"" + to_string<unsigned>(index) + "\"");
-  }
-}
-std::string MI::plink_write_bed_line(const MI::prob_vector &vec, const MI::annotations &annot, unsigned index) {
-  //need chr, rsid, pos, a1, a2 from annot
-  std::ostringstream o;
-  double p1 = 0.0, p2 = 0.0;//, p3 = 0.0;
-  unsigned res = 0;
-  std::string a1 = annot.get("a1", index);
-  std::string a2 = annot.get("a2", index);
-  for (unsigned i = 0; i < vec.size(); i += 2) {
-    p1 = vec.at(i);
-    p2 = vec.at(i+1);
-    /*p3 = 1.0 - p1 - p2;
-    if (p3 > 1.0) p3 = 1.0;
-    if (p3 < 0.0) p3 = 0.0;*/
-    res = standard_genotype_draw(p1, p2);
-    o << (3-res);
-  }
-  return o.str();
-}
-std::string MI::plink_write_bim_line(const MI::annotations &annot) {
-  std::ostringstream o;
-  std::vector<std::string> rsid = annot.get("rsid");
-  std::vector<std::string> pos = annot.get("pos");
-  std::vector<std::string> gpos = annot.get("gpos");
-  std::vector<std::string> chr = annot.get("chr");
-  std::vector<std::string> a1 = annot.get("a1");
-  std::vector<std::string> a2 = annot.get("a2");
-
-  if (rsid.empty() ||
-      a1.empty() ||
-      a2.empty())
-    throw std::domain_error("plink-write-bim-line: insufficient snp identifying information (rsid/a1/a2)");
-  for (unsigned i = 0; i < rsid.size(); ++i) {
-    o << (chr.empty() ? std::string("0") : chr.at(i)) << ' ' << rsid.at(i) << ' '
-      << (gpos.empty() ? std::string("0") : gpos.at(i)) << ' '
-      << (pos.empty() ? to_string<unsigned>(i+1) : pos.at(i)) << ' '
-      << a1.at(i) << ' ' << a2.at(i);
-    if (rsid.size() && i < (rsid.size()-1))
-      o << '\n';
-  }
-  return o.str();
-}
-std::string MI::plink_write_fam_line(const MI::annotations &annot) {
-  std::ostringstream o;
-  std::vector<std::string> fid = annot.get("fid");
-  std::vector<std::string> iid = annot.get("iid");
-  std::vector<std::string> pat = annot.get("pat");
-  std::vector<std::string> mat = annot.get("mat");
-  std::vector<std::string> sex = annot.get("sex");
-  //std::vector<std::string> pheno = annot.get("pheno");
-  if (iid.empty())
-    throw std::domain_error("plink-write-fam-line: no individual IDs present");
-  for (unsigned i = 0; i < iid.size(); ++i) {
-    o << (fid.empty() ? iid.at(i) : fid.at(i)) << ' ' << iid.at(i) << ' '
-      << (pat.empty() ? std::string("0") : pat.at(i)) << ' '
-      << (mat.empty() ? std::string("0") : mat.at(i)) << ' '
-      << (sex.empty() ? std::string("2") : sex.at(i)) << " -9";
-    if (iid.size() && i < (iid.size()-1))
-      o << '\n';
-  }
-  return o.str();
-}
-std::string MI::plink_write_pheno_line(const MI::annotations &annot) {
-  std::string requested_covariates = MI::parameters::get_parameter("plink-covar-names");
-  while (requested_covariates.find(",") != std::string::npos)
-    requested_covariates[requested_covariates.find(",")] = ' ';
-  std::string requested_phenotype = MI::parameters::get_parameter("plink-pheno-name");
-  std::ostringstream o;
-  o << "FID IID " << requested_phenotype << ' ' << requested_covariates << '\n';
-  std::vector<std::string> values;
-  std::istringstream strm1(requested_covariates);
-  std::string current_covariate = "";
-  std::vector<std::string> all_covariates;
-
-  while (strm1 >> current_covariate) {
-    all_covariates.push_back(current_covariate);
-  }
-  for (unsigned i = 0; i < annot.get_count_at("fid"); ++i) {
-    o << annot.get("fid", i) << ' ' << annot.get("iid", i) << " ";
-    o << annot.get(requested_phenotype, i);
-    for (std::vector<std::string>::const_iterator iter = all_covariates.begin(); iter != all_covariates.end(); ++iter) {
-      o << ' ' << annot.get(*iter, i);
-    }
-    if (i != annot.get_count_at("fid")-1) o << '\n';
-  }
-  return o.str();  
-}
-std::string MI::plink_format_command(std::string(*filename_generator)(unsigned, unsigned),
-				     unsigned draw) {
-  std::string command = "";
-  command += " --bed " + filename_generator(0, draw) + " --bim " + filename_generator(1, draw)
-    + " --fam " + filename_generator(2, draw);
-  if (MI::parameters::get_flag("plink-linear") &&
-      MI::parameters::get_flag("plink-logistic"))
-    throw std::domain_error("plink-format-command: either linear or logistic, not both, should be specified");
-  if (MI::parameters::get_flag("plink-linear"))
-    command += " --linear";
-  else if (MI::parameters::get_flag("plink-logistic"))
-    command += " --logistic";
-  else
-    throw std::domain_error("plink-format-command: neither linear nor logistic mode specified");
-  
-  command += " --pheno " + filename_generator(3, draw) + " --mpheno 1";
-  if (!MI::parameters::get_parameter("plink-covar-names").empty()) {
-    command += " --covar-name "
-    + MI::parameters::get_parameter("plink-covar-names");
-  }
-  if (!MI::parameters::get_parameter("plink-vif").empty()) {
-    command += " --vif " + MI::parameters::get_parameter("plink-vif");
-  }
-
-  command += " --out "
-    + MI::parameters::get_parameter("plink-out-prefix") + to_string<unsigned>(draw)
-    + MI::parameters::get_parameter("plink-out-suffix");
-  return command;  
-}
-bool MI::plink_process_results(const std::string &result_line,
-			       std::string &rsid,
-			       double &beta,
-			       double &stderr,
-			       std::string &effect_allele) {
-  if (result_line.find("TEST") != std::string::npos) return false;
-  std::istringstream strm1(result_line);
-  unsigned pos = 0, nmiss = 0;
-  std::string chr = "", test = "";
-  if (!(strm1 >> chr >> rsid >> pos >> effect_allele >> test >> nmiss >> beta >> stderr))
-    throw std::domain_error("plink-process-results: unable to parse line \"" + result_line + "\"");
-  if (MI::parameters::get_flag("plink-logistic")) {
-    beta = log(beta);
-  }
-  stderr = beta / stderr;
-  return true;
-}
-std::string MI::plink_extra_file_remover(unsigned draw) {
-  std::string res = "";
-  res += MI::parameters::get_parameter("plink-out-prefix") + to_string<unsigned>(draw)
-    + MI::parameters::get_parameter("plink-out-suffix") + ".log";
-  return res;
-}
+//#include "statgen-mi/custom_handlers_impute2.cc"
+//#include "statgen-mi/custom_handlers_snptest.cc"
+//#include "statgen-mi/custom_handlers_plink.cc"
+//#include "statgen-mi/custom_handlers_emmaxkin.cc"
+//#include "statgen-mi/custom_handlers_placeholder.cc"
